+++ examples/resized.rs
@@ -21,9 +21,17 @@ fn setup(mut commands: Commands) {
 
 fn update(mut q_term: Query<&mut Terminal>, time: Res<Time>) {
     let mut term = q_term.single_mut();
-    let size = (time.elapsed_secs().cos() * 20.0) as u32 % 20 + 10;
+    let size = (time.elapsed_secs().cos() * 20.0) as u32 % 20 + 15;
     term.resize([size, size]);
 
     term.clear();
-    term.put_string([0, 0].pivot(Pivot::Center), "Hello, World!");
+    term.put_string([0, 0].pivot(Pivot::TopLeft), "TopL");
+    term.put_string([0, 0].pivot(Pivot::TopCenter), "TopC");
+    term.put_string([0, 0].pivot(Pivot::TopRight), "TopR");
+    term.put_string([0, 0].pivot(Pivot::LeftCenter), "LefC");
+    term.put_string([0, 0].pivot(Pivot::Center), "C");
+    term.put_string([0, 0].pivot(Pivot::RightCenter), "RigC");
+    term.put_string([0, 0].pivot(Pivot::BottomLeft), "BotL");
+    term.put_string([0, 0].pivot(Pivot::BottomCenter), "BotC");
+    term.put_string([0, 0].pivot(Pivot::BottomRight), "BotR");
 }